lab 1 part 1: + stretch goal p.1 complete
got stuck working out the template. some of the reference pages are included in the comment block. I'm starting to get how to format _.template string, but not sure why the loop gets stuck after console.log(foodItems). Probably don't want to keep the extra assigned variables in the loop. keep it clean when code finalizes.