lab 2 progress: part 2
somewhat understand ajax calls and how the response gets stored. currently struggling with json cleaning process in part 2. should be able to wrap up lab2 once that is out of the way.
a few of the mapping functions were copied over from the previous assignments, but didn't get to test them yet